pyproject.toml
[tool.poetry]
name = "image-segmentation-blog-scripts"
version = "0.1.0"
description = ""
authors = ["Alex Thomas <alex.thomas@datatonic.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
tensorflow = "^2.8.1"
google-cloud-aiplatform = "^1.13.0"
Pillow = "9.1.1"

[tool.poetry.dev-dependencies]
pre-commit = "*"
black = "*"
isort = "*"
jupyterlab = "^3.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
